Refactor: Modify init structure and response
