+++ src/main/java/com/spring/blackcat/common/init/dto/InitResDto.java
@@ -1,0 +1,10 @@
+package com.spring.blackcat.common.init.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class InitResDto {
+    private String result;
+}
